Exempt Refaster templates from PMD checking
This is done because Refaster templates naturally violate certain PMD
checks that Refaster is itself meant to automatically fix, and the
currently solution of marking Refaster template classes with
`@SuppressWarnings("PMD")` is ugly.

To give an example, PMD complains if `list.size() == 0` is used instead
of `list.isEmpty()`, but an existing Refaster template that is designed
to automatically fix this sort of problem itself contains a call to
`list.size() == 0` so that code following this bad practice can be
caught.