Significantly rewrite Refaster Gradle plugin
It was decided to build upon gradle-errorprone-plugin after all to make
the implementation easier (and hopefully make integration between the
two plugins easier for users as well).

Currently, a detached dependency configuration is used to depend on a
very specific error-prone version that is known to work with Refaster.
This is done because I had lots of trouble getting Refaster to work
with other later versions of error-prone.

One reason that I did this rewrite was to allow multiple Refaster
templates to be used at a time.

Another reason I did this rewrite was to allow the tasks that are
introduced by the Refaster plugin to do more UP-TO-DATE checking and
be more incremental, to save user time in the long run.